+++ src/rlgym_learn_api/desc/project/project.py
@@ -1,8 +1,10 @@
+import os
+
 from pydantic import BaseModel
 
 
 class ProjectMetadata(BaseModel):
     name: str
     id: str
-    interpreter: str
+    interpreter: str | os.PathLike[str] | None
     created_at_version: str